# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KISS Messenger API Swagger",
    "description": "This is a small project demonstrating a limited feature set for an API that would serve a simple messenger app. This project was built as a take home project during the interview process at Guild Education.",
    "version": "1.0.0"
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "tags": [
    {
      "name": "Messages",
      "description": "Everything related to messages"
    }
  ],
  "paths": {
    "/messages": {
      "get": {
        "tags": ["Messages"],
        "summary": "Retrieve all messages sent to the recipient",
        "parameters": [
          {
            "$ref": "#/components/parameters/userId"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/daysElapsed"
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages sent to the recipient",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Messages"],
        "summary": "Send a message",
        "parameters": [
          {
            "$ref": "#/components/parameters/userId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipientId": {
                    "$ref": "#/components/schemas/UserId",
                    "example": 2
                  },
                  "text": {
                    "type": "string",
                    "example": "Hello, World!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of messages sent to the recipient",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/messages/{senderId}": {
      "get": {
        "tags": ["Messages"],
        "summary": "Retrieve all messages sent from the requested sender to recipient",
        "parameters": [
          {
            "name": "senderId",
            "description": "The User ID for the sender of the requested message set.",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserId"
            }
          },
          {
            "$ref": "#/components/parameters/userId"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/daysElapsed"
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages sent from the given sender to the reicipient",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Message": {
        "type": "object",
        "required": ["id", "metadata", "text", "recepientId", "senderId"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "dateSent": {
            "type": "string",
            "format": "date-time",
            "example": "2021-07-17T23:20:50.52Z"
          },
          "text": {
            "type": "string",
            "example": "Hello, World!"
          },
          "recepientId": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "senderId": {
            "type": "integer",
            "format": "int32",
            "example": 2
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "example": 501
          },
          "message": {
            "type": "string"
          }
        }
      },
      "UserId": {
        "type": "integer",
        "format": "int32",
        "example": 1
      }
    },
    "parameters": {
      "userId": {
        "name": "X-User-Id",
        "in": "header",
        "description": "The User ID for the recipient of the requested message set.",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/UserId"
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "limit the response list to this many items",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "default": 100
        }
      },
      "daysElapsed": {
        "name": "daysElapsed",
        "in": "query",
        "description": "limit the response list to items in the past n days",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "example": 30
        }
      }
    }
  }
}